Do not use VI mode
I have it found more and more annoying to use VI mode on my machines,
because the number of other machines I somethimes have to log into is
getting bigger. I do not want to have to maintain so many inputrc files
on so many machines, so I am switching to the default emacs mode, so
that it would become a reflex to use the emacs shortcuts. Also, I have
recently started learning Elixir and iex supports emacs mode, not VI
mode.

I still think VI shortcuts are superior.